pyproject.toml
[project]
name = "main"
version = "0.1.0"
description = "This is the main knowledge base."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ffmpeg>=1.4",
    "langchain>=0.3.27",
    "langchain-community>=0.3.28",
    "langchain-google-genai>=2.1.10",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.32",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
